Improve file structure
