+++ src/components/Dashboard/components/DashboardContent/components/BalanceChart/index.js
@@ -1,0 +1,19 @@
+import DoughnutChart from "../../../../../common/DoughnutChart";
+
+const BalanceChart = ({ incomesSum, expensesSum }) => {
+  const totalSum = incomesSum + Math.abs(expensesSum);
+  const incomePercentage = (incomesSum / totalSum) * 100;
+  const expensePercentage = (Math.abs(expensesSum) / totalSum) * 100;
+
+  return (
+    <DoughnutChart
+      data={{
+        labels: ["Incomes", "Expenses"],
+        chartData: [incomePercentage, expensePercentage],
+      }}
+      shouldShow={!!totalSum}
+    />
+  );
+};
+
+export default BalanceChart;
