+++ src/components/Dashboard/components/DashboardContent/styles.scss
@@ -1,4 +1,4 @@
-@import '../../variables.scss';
+@import '../../../../variables.scss';
 
 .dashboard-content {
   height: 100%;
